+++ src/routes/todos.ts
@@ -1,6 +1,7 @@
 import { Response, Router } from "express";
 import pool from "../db";
 import authorization from "../middlewares/authorization";
+import accessToProject from "../middlewares/privacyChecker";
 
 const todoRouter = Router();
 
@@ -28,35 +29,47 @@ todoRouter.get("/get-all-todos/:task_id", async (req: any, res: Response) => {
 });
 
 //create a todo
-todoRouter.post("/create-todo/:task_id", async (req: any, res: Response) => {
-  try {
-    const { task_id } = req.params;
-    type CreateTodo = {
-      description: string;
-      parent_todo: string;
-      index: number;
-      is_child: boolean;
-    };
-    const { description, parent_todo, index, is_child }: CreateTodo = req.body;
-    //find what index it needs to go ~ default index is the next one
-    const todos = await pool.query(
-      "SELECT * FROM todos WHERE task_id = $1 ORDER BY index ASC",
-      [task_id]
-    );
-    //is a parent todo ~ not indented
-    if (!is_child) {
-      const currentIndex = todos.rowCount;
+todoRouter.post(
+  "/create-todo/:project_id/:task_id",
+  [authorization, accessToProject],
+  async (req: any, res: Response) => {
+    try {
+      const { task_id } = req.params;
 
-      const createQuery = await pool.query(
-        "INSERT INTO todos (description, index, task_id) VALUES ($1, $2, $3)",
-        [description, currentIndex, task_id]
+      type CreateTodo = {
+        description: string;
+        parent_todo: string;
+        index: number;
+        is_child: boolean;
+      };
+      //parent todo is if we want to have a sub-todo(future feature)
+      //index are to keep track of what position a sub-todo would be in
+      const { description, parent_todo, index, is_child }: CreateTodo =
+        req.body;
+
+      const editingStatus: boolean = req.editingStatus;
+      if (!editingStatus) return res.send(undefined);
+
+      //find what index it needs to go ~ default index is the next one
+      const todos = await pool.query(
+        "SELECT * FROM todos WHERE task_id = $1 ORDER BY index ASC",
+        [task_id]
       );
-      res.send("added a new todo");
+      //is a parent todo ~ not indented
+      if (!is_child) {
+        const currentIndex = todos.rowCount;
+
+        await pool.query(
+          "INSERT INTO todos (description, index, task_id) VALUES ($1, $2, $3)",
+          [description, currentIndex, task_id]
+        );
+        res.send({ success: true, message: "added a new todo" });
+      }
+    } catch (error) {
+      console.error(error);
     }
-  } catch (error) {
-    console.error(error);
   }
-});
+);
 
 //toggle todo completion
 todoRouter.put(
@@ -103,18 +116,19 @@ todoRouter.put(
 
 //delete todo
 todoRouter.delete(
-  "/delete-todo/:todo_id",
-  authorization,
+  "/delete-todo/:project_id/:todo_id",
+  [authorization, accessToProject],
   async (req: any, res: Response) => {
     try {
-      const user_id = req.user;
+      const editingStatus: boolean = req.editingStatus;
+      if (!editingStatus) return res.send(undefined);
 
       const { todo_id } = req.params;
 
       const query = await pool.query("DELETE FROM todos WHERE todo_id = $1 ", [
         todo_id,
       ]);
-      res.send("Successfully Deleted ");
+      res.send({ success: true, message: "Successfully Deleted " });
     } catch (error) {
       console.error(error);
     }
